requirements.txt
Flask
uuid
gradio
requests
langchain
langchain-community
langchain-core
langchain-openai
reportlab
pytesseract
SpeechRecognition
fastembed
python-dotenv
pypdf
faiss-cpu